+++ lunchmenu/helper/__init__.py
@@ -1,2 +1,3 @@
 from helper import Params
 from helper import Zomato
+from helper import Websites